requirements.txt
request
spacy
newspaper
